Rename utils.py to db.py
